+++ src/builder/command.rs
@@ -4391,16 +4391,8 @@ To change `help`s short, call `cmd.arg(Arg::new(\"help\")...)`.",
         } else {
             self.version.or(self.long_version).unwrap_or("")
         };
-        if let Some(bn) = self.bin_name.as_ref() {
-            if bn.contains(' ') {
-                // In case we're dealing with subcommands i.e. git mv is translated to git-mv
-                format!("{} {}\n", bn.replace(' ', "-"), ver)
-            } else {
-                format!("{} {}\n", &self.name[..], ver)
-            }
-        } else {
-            format!("{} {}\n", &self.name[..], ver)
-        }
+        let display_name = self.get_display_name().unwrap_or_else(|| self.get_name());
+        format!("{} {}\n", display_name, ver)
     }
 
     pub(crate) fn format_group(&self, g: &Id) -> String {